refactore bus module, add exception handling

+++ bus/src/main/java/com/skopinau/bus/repository/BusRepository.java
@@ -1,7 +1,7 @@
 package com.skopinau.bus.repository;
 
-import com.skopinau.bus.model.entity.Bus;
+import com.skopinau.bus.entity.Bus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusRepository extends JpaRepository<Bus, Integer> {
+public interface BusRepository extends JpaRepository<Bus, Long> {
 }
